requirements.txt
accelerate==1.6.0
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
bitsandbytes==0.45.5
black==25.1.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flake8==7.2.0
fsspec==2025.3.2
gdown==5.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
isort==6.0.1
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
preshed==3.0.9
psutil==7.0.0
pycodestyle==2.13.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
PySocks==1.7.1
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-crf==0.7.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.5
safetensors==0.5.3
setuptools==80.3.1
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
watchfiles==1.0.5
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
